+++ pkg/apis/lbaas/v2/common.go
@@ -1,0 +1,7 @@
+package v2
+
+// commonRequestBody is used to optionally omit the `State` field on create and update
+// by embedding it to the request in the FilterAPIRequestBody hook
+type commonRequestBody struct {
+	State string `json:"state,omitempty"`
+}